# Elimination of abstract type parameters and fields -> performance improvement