--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,5 @@
 docs/build
 *.DS_Store
 examples/.ipynb_checkpoints/
+
+*.mem