--- a/src/boundary_conditions.jl
+++ b/src/boundary_conditions.jl
@@ -8,31 +8,30 @@ struct Gradient <: BCType end
 struct Value <: BCType end
 
 "Return c₁, where c₁ = c₀ + ∂c/∂z * (z₁ - z₀) = c₀ + ∂c/∂z * Δz."
-linear_extrap(c₀, ∂c∂z, Δz) = c₀ + ∂c∂z * Δz
+@inline linear_extrap(c₀, ∂c∂z, Δz) = c₀ + ∂c∂z * Δz
 
-abstract type AbstractBoundaryCondition end
-
-struct BoundaryCondition{C, T} <: AbstractBoundaryCondition
+mutable struct BoundaryCondition{C, T}
     condition :: T
 end
 
+function Base.convert(::Type{BoundaryCondition{C, T1}}, bc::BoundaryCondition{C}) where {C, T1}
+    return BoundaryCondition{C, T1}(convert(T1, bc.condition))
+end
+
 BoundaryCondition(bctype, bc) = BoundaryCondition{bctype, typeof(bc)}(bc)
+
 GradientBoundaryCondition(bc) = BoundaryCondition(Gradient, bc)
 FluxBoundaryCondition(bc) = BoundaryCondition(Flux, bc)
 ValueBoundaryCondition(bc) = BoundaryCondition(Value, bc)
 
-struct ZeroFluxBoundaryCondition <: AbstractBoundaryCondition end
-
 const BC = BoundaryCondition
-const ZFBC = ZeroFluxBoundaryCondition
 
-getbc(model, ::ZFBC) = 0
-getbc(model, bc::BC{C, <:Number}) where C = bc.condition
-getbc(model, bc::BC{C, <:Function}) where C = bc.condition(model)
+@inline getbc(model, bc::BC{C, <:Number}) where C = bc.condition
+@inline getbc(model, bc::BC{C, <:Function}) where C = bc.condition(model)
 
-gradient(bc::BC{<:Gradient}, model, args...) = getbc(model, bc)
-gradient(bc::BC{<:Flux},     model, κ, args...) = -getbc(model, bc) / κ
-gradient(bc::BC{<:Value},    model, κ, cᴺ, Δf, args...) = 2 * (getbc(model, bc) - cᴺ) / Δf
+@inline gradient(bc::BC{<:Gradient}, model, args...) = getbc(model, bc)
+@inline gradient(bc::BC{<:Flux},     model, κ, args...) = -getbc(model, bc) / κ
+@inline gradient(bc::BC{<:Value},    model, κ, cᴺ, Δf, args...) = 2 * (getbc(model, bc) - cᴺ) / Δf
 
 """
     fill_bottom_ghost_cell!(c, κ, model, bc)
@@ -41,7 +40,7 @@ Update the bottom ghost cell of c given the boundary condition `bc`, `model`, an
 diffusivity `kappa`. `kappa` is used only if a flux boundary condition
 is specified.
 """
-function fill_bottom_ghost_cell!(bc, c, κ, model)
+@inline function fill_bottom_ghost_cell!(bc, c, κ, model)
     @inbounds c[0] = linear_extrap(c[1], gradient(bc, model, κ, c[1], -Δf(c, 1)), -Δc(c, 1))
     return nothing
 end
@@ -52,31 +51,47 @@ end
 Update the top ghost cell of `c` given boundary condition `bc`, `model`, and
 diffusivity `kappa`
 """
-function fill_top_ghost_cell!(bc, c, κ, model, N)
+@inline function fill_top_ghost_cell!(bc, c, κ, model, N)
     @inbounds c[N+1] = linear_extrap(c[N], gradient(bc, model, κ, c[N], Δf(c, N)), Δc(c, N+1))
     return nothing
-end
-
-fill_bottom_ghost_cell!(::ZFBC, c, args...) = @inbounds c[0] = c[1]
-fill_top_ghost_cell!(::ZFBC, c, κ, model, N) = @inbounds c[N+1] = c[N]
-
-mutable struct FieldBoundaryConditions
-    bottom :: AbstractBoundaryCondition
-    top    :: AbstractBoundaryCondition
 end
 
 """
     FieldBoundaryConditions(; top=TopBC, bottom=BottomBC)
 
 Create an instance of `FieldBoundaryConditions` with `top` and `bottom`
-boundary conditions.
 """
-function FieldBoundaryConditions(; bottom = ZeroFluxBoundaryCondition(),
-                                      top = ZeroFluxBoundaryCondition())
-    FieldBoundaryConditions(bottom, top)
+mutable struct FieldBoundaryConditions{B, T}
+    bottom :: B
+    top    :: T
 end
 
-ZeroFluxBoundaryConditions() = FieldBoundaryConditions()
+"""
+    ZeroFluxBoundaryConditions(T=Float64)
+
+Construct `FieldBoundaryConditions` with a zero `FluxBoundaryCondition`
+at top and bottom.
+"""
+function ZeroFluxBoundaryConditions(T=Float64)
+    FieldBoundaryConditions(
+        FluxBoundaryCondition(-zero(T)),
+        FluxBoundaryCondition(-zero(T))
+    )
+end
+
+"""
+    DefaultBoundaryConditions(T=Float64)
+
+Returns default oceanic boundary conditions: a zero
+`GradientBoundaryCondition` on bottom and a zero
+`FluxBoundaryCondition` on top.
+"""
+function DefaultBoundaryConditions(T=Float64)
+    return FieldBoundaryConditions(
+                GradientBoundaryCondition(-zero(T)),
+                FluxBoundaryCondition(-zero(T))
+                )
+end
 
 """
     set_bcs!(model; bcspecs...)