--- a/test/diffusiontests.jl
+++ b/test/diffusiontests.jl
@@ -1,4 +1,4 @@
-#
+
 # Tests for the Diffusion module
 #