On current master:

```
$ julia benchmarks/atomic_benchmarks_kpp.jl 

Atomic timestepping benchmarking of default configuration...
N:  128, forward Euler
                 prepare:  33.290 μs (1190 allocations: 18.59 KiB)
       calc explicit rhs:  53.316 μs (1920 allocations: 30.00 KiB)
         update solution:  913.211 ns (0 allocations: 0 bytes)
                    tick:  2.598 ns (0 allocations: 0 bytes)
N:  128, backward Euler
                 prepare:  30.387 μs (1190 allocations: 18.59 KiB)
       calc implicit rhs:  28.536 μs (912 allocations: 14.25 KiB)
      calc diffusive lhs:  158.460 μs (6104 allocations: 95.38 KiB)
         update solution:  6.795 μs (8 allocations: 384 bytes)
                    tick:  2.599 ns (0 allocations: 0 bytes)

Atomic timestepping benchmarking of configuration with one constant boundary condition...
N:  128, forward Euler
                 prepare:  31.204 μs (1189 allocations: 18.58 KiB)
       calc explicit rhs:  52.495 μs (1920 allocations: 30.00 KiB)
         update solution:  913.395 ns (0 allocations: 0 bytes)
                    tick:  2.598 ns (0 allocations: 0 bytes)
N:  128, backward Euler
                 prepare:  35.908 μs (1189 allocations: 18.58 KiB)
       calc implicit rhs:  28.480 μs (912 allocations: 14.25 KiB)
      calc diffusive lhs:  154.553 μs (6104 allocations: 95.38 KiB)
         update solution:  7.136 μs (8 allocations: 384 bytes)
                    tick:  2.598 ns (0 allocations: 0 bytes)
```

With changes in this PR:

```
$ julia benchmarks/atomic_benchmarks_kpp.jl 

Atomic timestepping benchmarking of default configuration...
N:  128, forward Euler
                  update:  10.914 μs (8 allocations: 128 bytes)
       calc explicit rhs:  9.274 μs (0 allocations: 0 bytes)
         update solution:  81.651 ns (0 allocations: 0 bytes)
                    tick:  2.599 ns (0 allocations: 0 bytes)
N:  128, backward Euler
                  update:  10.529 μs (8 allocations: 128 bytes)
       calc implicit rhs:  136.419 ns (0 allocations: 0 bytes)
      calc diffusive lhs:  36.375 μs (0 allocations: 0 bytes)
         update solution:  6.787 μs (8 allocations: 384 bytes)
                    tick:  2.934 ns (0 allocations: 0 bytes)

Atomic timestepping benchmarking of configuration with one constant boundary condition...
N:  128, forward Euler
                  update:  12.784 μs (8 allocations: 128 bytes)
       calc explicit rhs:  77.516 μs (0 allocations: 0 bytes)
         update solution:  82.050 ns (0 allocations: 0 bytes)
                    tick:  2.599 ns (0 allocations: 0 bytes)
N:  128, backward Euler
                  update:  13.389 μs (8 allocations: 128 bytes)
       calc implicit rhs:  2.140 μs (0 allocations: 0 bytes)
      calc diffusive lhs:  181.456 μs (0 allocations: 0 bytes)
         update solution:  6.782 μs (8 allocations: 384 bytes)
                    tick:  2.675 ns (0 allocations: 0 bytes)
```

Thus:

1. With default boundary conditions (constant flux / gradient of 0), constant propagation eliminates much of the cost of evaluating KPP. This feature will rarely be used but is good to see the compiler is doing its job.

2.  There's less memory allocation.

3. The bottlenecks are now correctly the steps in which the K-profile is calculated --- `calc_explicit_rhs!` and `calc_diffusive_lhs`.

4. Performance is improved, though the improvement isn't huge when the boundary conditions are non-constant. Possibly more improvements are to be had in the calculation of the KPP K-profile.

cc @vchuravy